% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{setioalbgrid}
\alias{setioalbgrid}
\title{Sets a grid of SS3 model runs for Indian Ocean albacore}
\usage{
setioalbgrid(
  sce,
  dir = paste0("grid_", format(Sys.time(), "\%Y\%m\%d")),
  base = system.file("ext-data/sa/2016", package = "ioalbmse"),
  name = "abt",
  from = 1,
  write = TRUE,
  delete = TRUE
)
}
\arguments{
\item{sce}{[TODO:description]}

\item{name}{[TODO:description]}

\item{dir=paste0}{[TODO:description]}

\item{grid_}{[TODO:description]}

\item{format(Sys.time}{[TODO:description]}

\item{Y%m%d}{[TODO:description]}

\item{base=system.file}{[TODO:description]}

\item{ext-data/sa/2016}{[TODO:description]}

\item{package}{[TODO:description]}

\item{from=1}{[TODO:description]}

\item{write=TRUE}{[TODO:description]}

\item{delete=TRUE}{[TODO:description]}
}
\value{
[TODO:description]
}
\description{
Sets a grid of SS3 model runs for Indian Ocean albacore
}
